+++ packages/mix/lib/mix.dart
@@ -40,6 +40,7 @@ export 'src/core/prop_source.dart';
 export 'src/core/providers/style_provider.dart';
 export 'src/core/providers/style_spec_provider.dart';
 export 'src/core/providers/widget_state_provider.dart';
+export 'src/core/providers/widget_state_style_override.dart';
 export 'src/core/spec.dart';
 export 'src/core/style.dart' hide StyleElement;
 export 'src/core/style_builder.dart';